Nuevas logicas y botones, se corrgie Javascript

+++ operaciones/views.py
@@ -1,4 +1,3 @@
-from turtledemo.penrose import f
 from django.db.models import Sum
 from django.http import JsonResponse
 from django.template.loader import render_to_string
@@ -6,7 +5,7 @@ from django.views.generic.edit import CreateView
 from django_tables2 import SingleTableView
 from django.utils import timezone
 from django.shortcuts import render
-from .forms import MovimientoForm, ItemForm
+from .forms import MovimientoForm, ItemForm, MovimientoSearchForm
 from .tables import MovimientoTable, ItemTable, InventariorealTable
 from .models import Bodega, Item, Movimiento
 
@@ -19,6 +18,7 @@ def inventariotr(request):
     return render(request, 'inventariotr.html')
 
 
+# Movimientos. (Inventario Real Aux Admin)
 def mover_item(request):
     usuario = request.user
     if request.method == 'POST':
@@ -28,40 +28,36 @@ def mover_item(request):
             cantidad = form.cleaned_data['cantidad']
             bodega_destino = form.cleaned_data['bodega_destino']
             if cantidad > 0:
-                if cantidad <= item.kilos_netos:
-                    if item.bodega != bodega_destino:
-                        if bodega_destino:
-                            try:
-                                item_destino = Item.objects.get(numero_item=item.numero_item, bodega=bodega_destino)
-                                item_destino.kilos_netos += cantidad
-                                item_destino.save()
-                            except Item.DoesNotExist:
-                                Item.objects.create(numero_item=item.numero_item, kilos_netos=cantidad,
-                                                    bodega=bodega_destino, fruta=item.fruta,
-                                                    tipo_negociacion=item.tipo_negociacion, user=usuario)
-                        item.kilos_netos -= cantidad
-                        item.save()
-                        movimiento = Movimiento(item_historico=item.numero_item, cantidad=cantidad,
-                                                bodega_origen=item.bodega,
-                                                bodega_destino=bodega_destino, fruta=item.fruta,
-                                                t_negociacion=item.tipo_negociacion, user=usuario)
-                        movimiento.save()
-                        if item.kilos_netos == 0:
-                            item.delete()
-
-                        return JsonResponse({'success': True})
+                if cantidad <= item.kilos_netos and item.bodega != bodega_destino:
+                    try:
+                        item_destino = Item.objects.get(numero_item=item.numero_item, bodega=bodega_destino)
+                        item_destino.kilos_netos += cantidad
+                        item_destino.save()
+                    except Item.DoesNotExist:
+                        Item.objects.create(numero_item=item.numero_item, kilos_netos=cantidad,
+                                            bodega=bodega_destino, fruta=item.fruta,
+                                            tipo_negociacion=item.tipo_negociacion, user=usuario)
+                    item.kilos_netos -= cantidad
+                    item.save()
+                    movimiento = Movimiento(item_historico=item.numero_item, cantidad=cantidad,
+                                            bodega_origen=item.bodega,
+                                            bodega_destino=bodega_destino, fruta=item.fruta,
+                                            t_negociacion=item.tipo_negociacion, user=usuario)
+                    movimiento.save()
+                    if item.kilos_netos == 0:
+                        item.delete()
+                    return JsonResponse({'success': True})
+                elif item.bodega == bodega_destino:
+                    error_msg = f"La bodega de origen -> {item.bodega}, es igual a la bodega destino -> {bodega_destino}"
+                    return JsonResponse({'success': False, 'error': error_msg})
                 else:
-                    f.errors.as_json('cantidad',
-                                     f"No hay suficiente stock disponible para dar salida a {cantidad} kilos netos.")
-                    return JsonResponse({'success': False, 'error': str(form.errors['cantidad'])})
-
+                    error_msg = f"No hay suficiente stock disponible para dar salida a {cantidad} kilos netos. Kilos Netos disponibles: {item.kilos_netos}"
+                    return JsonResponse({'success': False, 'error': error_msg})
             else:
-                f.errors.as_json('cantidad', "La cantidad de kilos netos debe ser mayor que 0.")
-                return JsonResponse({'success': False, 'error': str(form.errors['cantidad'])})
-
+                error_msg = "La cantidad de kilos netos debe ser mayor que 0."
+                return JsonResponse({'success': False, 'error': error_msg})
     else:
         form = MovimientoForm()
-
     for bodega in Bodega.objects.all():
         items_bodega = Item.objects.filter(bodega=bodega).distinct('numero_item')
         for item in items_bodega:
@@ -69,7 +65,6 @@ def mover_item(request):
                 total=Sum('kilos_netos'))['total']
             item.kilos_netos = cantidad_total
             item.save()
-
     items = Item.objects.exclude(bodega__nombre="Calidad").filter(kilos_netos__gt=0, bodega__isnull=False)
     table = InventariorealTable(items)
     return render(request, 'mover_item.html', {'form': form, 'table': table})
@@ -80,6 +75,21 @@ class MovimientoListView(SingleTableView):
     model = Movimiento
     table_class = MovimientoTable
     template_name = 'historico.html'
+    form_class = MovimientoSearchForm
+
+    def get_queryset(self):
+        queryset = super().get_queryset()
+        form = self.form_class(self.request.GET)
+        if form.is_valid():
+            item_busqueda = form.cleaned_data.get('item_busqueda')
+            if item_busqueda:
+                queryset = queryset.filter(item_historico__icontains=item_busqueda)
+        return queryset
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context['item_busqueda'] = self.form_class(self.request.GET)
+        return context
 
 
 # Mostrar Tabla Recibo - Bodega Recibo (Inventario Real)
@@ -108,6 +118,34 @@ class ItemCreateView(CreateView):
         return initial
 
     def form_valid(self, form):
+        numero_item = form.cleaned_data['numero_item']
+        fruta = form.cleaned_data['fruta']
+        bodega = form.cleaned_data['bodega']
+        kilos_netos = form.cleaned_data['kilos_netos']
+
+        if Item.objects.filter(numero_item=numero_item).exclude(fruta=fruta).exists():
+            error_msg = f'Ya existe este Item {numero_item}, con fruta diferente.'
+            return JsonResponse({'success': False, 'error': error_msg})
+
+        # Buscar el item existente en la misma bodega con el mismo numero_item y fruta
+        try:
+            existing_item = Item.objects.get(numero_item=numero_item, fruta=fruta, bodega=bodega)
+            existing_item.kilos_netos += kilos_netos  # Sumar los kilos_netos
+            existing_item.save()
+            Movimiento.objects.create(
+                item_historico=existing_item.numero_item,
+                cantidad=existing_item.kilos_netos,
+                bodega_origen=existing_item.bodega,
+                bodega_destino=existing_item.bodega,
+                fruta=existing_item.fruta.nombre_fruta,
+                t_negociacion=existing_item.tipo_negociacion,
+                fecha=timezone.now(),
+                user=existing_item.user
+            )
+            return JsonResponse({'success': True})
+        except Item.DoesNotExist:
+            pass
+
         form.instance.user = self.request.user
         item = form.save()
         Movimiento.objects.create(
